requirements.txt
flask
flask_sqlalchemy
flask_migrate
psycopg2-binary
python-dotenv
aiohttp
httpx
beautifulsoup4
readability-lxml
lxml[html_clean]
openai
pycountry
asyncio
trafilatura
requests
pandas
openpyxl
apscheduler
mysql-connector-python
amadeus
Flask[async]
Jinja2
Mako==1.2.4
MarkupSafe==2.1.2
flask-bcrypt
flask-login